+++ src/components/Heading/Heading.stories.tsx
@@ -68,48 +68,114 @@ export const WithControls: Story = {
 
 // The following are additional Heading example Stories.
 export const DefaultStyles: Story = {
+  name: "New Defaults",
   render: () => (
     <>
-      <Heading id="heading-one" level="one" text="<h1> Heading 1 - 'one'" />
-      <Heading id="heading-two" level="two" text="<h2> Heading 2 - 'two'" />
-      <Heading
-        id="heading-three"
-        level="three"
-        text="<h3> Heading 3 - 'three'"
-      />
-      <Heading id="heading-four" level="four" text="<h4> Heading 4 - 'four'" />
-      <Heading id="heading-five" level="five" text="<h5> Heading 5 - 'five'" />
-      <Heading id="heading-six" level="six" text="<h6> Heading 6 - 'six'" />
+      <Heading id="heading-h1" level="h1" text="(h1) Lorem ipsum dolor" />
+      <Heading id="heading-h2" level="h2" text="(h2) Lorem ipsum dolor" />
+      <Heading id="heading-h3" level="h3" text="(h3) Lorem ipsum dolor" />
+      <Heading id="heading-h4" level="h4" text="(h4) Lorem ipsum dolor" />
+      <Heading id="heading-h5" level="h5" text="(h5) Lorem ipsum dolor" />
+      <Heading id="heading-h6" level="h6" text="(h6) Lorem ipsum dolor" />
     </>
   ),
 };
 
 export const SizeStyles: Story = {
+  name: "New Sizes",
+  render: () => (
+    <>
+      <Heading
+        id="heading-display1"
+        level="h1"
+        size="display1"
+        text="(display1) Lorem ipsum dolor"
+      />
+      <Heading
+        id="heading-heading1"
+        level="h1"
+        size="heading1"
+        text="(heading1) Lorem ipsum dolor"
+      />
+      <Heading
+        id="heading-heading2"
+        level="h1"
+        size="heading2"
+        text="(heading2) Lorem ipsum dolor"
+      />
+      <Heading
+        id="heading-heading3"
+        level="h1"
+        size="heading3"
+        text="(heading3) Lorem ipsum dolor"
+      />
+      <Heading
+        id="heading-heading4"
+        level="h1"
+        size="heading4"
+        text="(heading4) Lorem ipsum dolor"
+      />
+      <Heading
+        id="heading-heading5"
+        level="h1"
+        size="heading5"
+        text="(heading5) Lorem ipsum dolor"
+      />
+      <Heading
+        id="heading-heading6"
+        level="h1"
+        size="heading6"
+        text="(heading6) Lorem ipsum dolor"
+      />
+    </>
+  ),
+};
+
+export const DefaultStylesDeprecated: Story = {
+  name: "Deprecated Defaults",
+  render: () => (
+    <>
+      <Heading id="heading-one" level="one" text="(one) Lorem ipsum dolor" />
+      <Heading id="heading-two" level="two" text="(two) Lorem ipsum dolor" />
+      <Heading
+        id="heading-three"
+        level="three"
+        text="(three) Lorem ipsum dolor"
+      />
+      <Heading id="heading-four" level="four" text="(four) Lorem ipsum dolor" />
+      <Heading id="heading-five" level="five" text="(five) Lorem ipsum dolor" />
+      <Heading id="heading-six" level="six" text="(six) Lorem ipsum dolor" />
+    </>
+  ),
+};
+
+export const SizeStylesDeprecated: Story = {
+  name: "Deprecated Sizes",
   render: () => (
     <>
       <Heading
         id="heading-primary"
         level="one"
         size="primary"
-        text="<h1> 'primary'"
+        text="(primary) Lorem ipsum dolor"
       />
       <Heading
         id="heading-secondary"
         level="one"
         size="secondary"
-        text="<h1> 'secondary'"
+        text="(secondary) Lorem ipsum dolor"
       />
       <Heading
         id="heading-tertiary"
         level="one"
         size="tertiary"
-        text="<h1> 'tertiary'"
+        text="(tertiary) Lorem ipsum dolor"
       />
       <Heading
         id="heading-callout"
         level="one"
         size="callout"
-        text="<h1> 'callout'"
+        text="(callout) Lorem ipsum dolor"
       />
     </>
   ),
@@ -117,7 +183,7 @@ export const SizeStyles: Story = {
 
 export const BoldText: Story = {
   render: () => (
-    <Heading id="headingWithBold" level="one">
+    <Heading id="headingWithBold" level="h2">
       <>
         Heading with a <b>Bold</b> Word
       </>
@@ -131,12 +197,12 @@ export const Links: Story = {
     <>
       <Heading
         id="heading-with-link-url"
-        level="two"
+        level="h2"
         text="Heading with URL Props"
         url="http://apple.com"
         urlClass="special-link"
       />
-      <Heading id="heading-with-link-child">
+      <Heading id="heading-with-link-child" level="h2">
         <>
           Heading with a Word <a href="#hello">Link</a>
         </>