+++ src/components/Text/Text.stories.tsx
@@ -2,6 +2,9 @@ import type { Meta, StoryObj } from "@storybook/react";
 import { withDesign } from "storybook-addon-designs";
 
 import Text, { textSizesArray } from "./Text";
+import Heading from "../Heading/Heading";
+import Image from "../Image/Image";
+import { Box } from "@chakra-ui/react";
 
 const meta: Meta<typeof Text> = {
   title: "Components/Typography & Styles/Text",
@@ -71,6 +74,99 @@ export const WithControls: Story = {
         games have also been released: Animal Crossing: Happy Home Designer for
         Nintendo 3DS, Animal Crossing: Amiibo Festival for Wii U, and Animal
         Crossing: Pocket Camp for mobile devices.
+      </Text>
+    </>
+  ),
+};
+
+// The following are additional Heading example Stories.
+export const SizeStyles: Story = {
+  render: () => (
+    <>
+      <Text size="subtitle1">Subtitle 1</Text>
+      <Text size="subtitle2">Subtitle 2</Text>
+      <Text size="overline1">Overline 1</Text>
+      <Text size="overline2">Overline 2</Text>
+      <Text size="default">Default text</Text>
+      <Text size="caption">Caption text</Text>
+      <Text size="tag">Tag text</Text>
+      <Text size="mini">Mini text (deprecated)</Text>
+    </>
+  ),
+};
+export const DefaultExample: Story = {
+  render: () => (
+    <>
+      <Text>
+        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent
+        commodo cursus magna, vel scelerisque nisl consectetur et. Donec id elit
+        non mi porta gravida at eget metus. Aenean lacinia bibendum nulla sed
+        consectetur. Vestibulum id ligula porta felis euismod semper. Cras justo
+        odio, dapibus ac facilisis in, egestas eget quam.
+      </Text>
+    </>
+  ),
+};
+export const CaptionExample: Story = {
+  render: () => (
+    <>
+      <Box width="fit-content">
+        <Image src="//placekitten.com/400/400" size="medium" />
+        <Text size="caption">This is a kitten</Text>
+      </Box>
+    </>
+  ),
+};
+export const TagExample: Story = {
+  render: () => (
+    <>
+      <Text size="tag">Tag text</Text>
+    </>
+  ),
+};
+export const MiniExample: Story = {
+  render: () => (
+    <>
+      <Text size="mini">Mini text</Text>
+    </>
+  ),
+};
+export const Overline1Example: Story = {
+  render: () => (
+    <>
+      <Text size="overline1" noSpace>
+        Category
+      </Text>
+      <Heading size="heading2">Lorem ipsum dolor</Heading>
+    </>
+  ),
+};
+export const Overline2Example: Story = {
+  render: () => (
+    <>
+      <Text size="overline2" noSpace>
+        Category
+      </Text>
+      <Heading size="heading5">Lorem ipsum dolor</Heading>
+    </>
+  ),
+};
+export const Subtitle1Example: Story = {
+  render: () => (
+    <>
+      <Heading size="heading2">Lorem ipsum dolor</Heading>
+      <Text size="subtitle1" noSpace>
+        Nullam id dolor id nibh ultricies vehicula ut id elit.
+      </Text>
+    </>
+  ),
+};
+export const Subtitle2Example: Story = {
+  render: () => (
+    <>
+      <Heading size="heading5">Lorem ipsum dolor</Heading>
+      <Text size="subtitle2" noSpace>
+        Nullam id dolor id nibh ultricies vehicula ut id elit.
       </Text>
     </>
   ),