+++ src/theme/components/text.ts
@@ -11,13 +11,35 @@ interface TextBaseStyle {
 const variants = {
   default: {},
   caption: {
-    fontSize: "text.caption",
+    fontSize: "desktop.caption",
   },
   tag: {
-    fontSize: "text.tag",
+    fontSize: "desktop.tag",
   },
   mini: {
     fontSize: "text.mini",
+  },
+  overline1: {
+    fontSize: "desktop.overline.overline1",
+    fontWeight: "text.overline",
+    lineHeight: 1.35,
+    textTransform: "uppercase",
+  },
+  overline2: {
+    fontSize: "desktop.overline.overline2",
+    fontWeight: "text.overline",
+    lineHeight: 1.4,
+    textTransform: "uppercase",
+  },
+  subtitle1: {
+    fontSize: "desktop.subtitle.subtitle1",
+    fontWeight: "text.subtitle",
+    lineHeight: 1.25,
+  },
+  subtitle2: {
+    fontSize: "desktop.subtitle.subtitle2",
+    fontWeight: "text.subtitle",
+    lineHeight: 1.3,
   },
 };
 const Text = {