+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Updates the `Button` component to increase the internal spacing between the button text and an icon.
+- Updates the existing `heading` typography styles to add new default styles options for native heading elements, including `h1` thru `h6`.
+- Updates the existing `heading` typography styles to add new size options, including `display1` and `heading1` thru `heading6`.
+- Updates the existing `text` typography styles to add `overline` and `subtitle` options.
+- Updates the `Typography` docs to include the new options added to the `heading` and `text` typography styles.
+- Updates the `Design Tokens` docs to include the new tokens added to the `heading` and `text` typography styles.
 
 ## 1.6.0 (June 8, 2023)